@@ -1,11 +1,11 @@
 import { browser, by, element } from 'protractor';
 
 export class AppPage {
-  navigateTo(): Promise<unknown> {
-    return browser.get(browser.baseUrl) as Promise<unknown>;
+  navigateTo() {
+    return browser.get(browser.baseUrl) as Promise<any>;
   }
 
-  getTitleText(): Promise<string> {
+  getTitleText() {
     return element(by.css('app-root .content span')).getText() as Promise<string>;
   }
 }
