@@ -20,7 +20,14 @@ const routes: Routes = [
       .then(mod => mod.CheckoutModule), data: { breadcrumb: 'Checkout' }
   },
   {
-    path: 'account', loadChildren: () => import('./account/account.module')
+    path: 'orders',
+    canActivate: [AuthGuard],
+    loadChildren: () => import('./orders/orders.module')
+      .then(mod => mod.OrdersModule), data: { breadcrumb: 'Orders' }
+  },
+  {
+    path: 'account',
+    loadChildren: () => import('./account/account.module')
       .then(mod => mod.AccountModule), data: { breadcrumb: { skip: true } }
   },
   { path: '**', redirectTo: 'not-found', pathMatch: 'full' }
